Add support for a focal detection threshold
The focal detection threshold is a confidence rating less than the detection threshold which is used to trigger focus without taking a picture or triggering a video recording to start.
This adjusted focus may help to improve the confidence rating of subsequent detections or provide better autofocus while recording due to an object triggering recording with a high confidence but then subsequently being rated with a lower confidence.